Separated old code from new code and redistributed Auto functions to their own class
